shift code from messageCTA to calendarEventButton
